requirements.txt
discord.py
python-dotenv
pymongo # or database of choice